chore: devide test file into module folder
